requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
docker==7.1.0
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
Flask-SocketIO==5.5.1
eventlet==0.38.2
flask_login==0.6.3
asyncpg==0.30.0
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
python-dotenv==1.0.1